requirements.txt
python-dotenv
requests
rich
langchain-ollama
openai
anthropic
tavily-python
ollama
langchain-ollama